Improve consistency among resources
