+++ libvirt/resource_libvirt_domain.go
@@ -6,8 +6,6 @@ import (
 	"log"
 	"net"
 	"net/url"
-	"sort"
-	"strconv"
 	"strings"
 	"time"
 
@@ -949,551 +947,6 @@ func resourceLibvirtDomainDelete(d *schema.ResourceData, meta interface{}) error
 			}
 		} else {
 			return fmt.Errorf("Couldn't undefine libvirt domain with flags: %s", err)
-		}
-	}
-
-	return nil
-}
-
-func newDiskForCloudInit(virConn *libvirt.Connect, volumeKey string) (libvirtxml.DomainDisk, error) {
-	disk := libvirtxml.DomainDisk{
-		Device: "cdrom",
-		Target: &libvirtxml.DomainDiskTarget{
-			Dev: "hda",
-			Bus: "ide",
-		},
-		Driver: &libvirtxml.DomainDiskDriver{
-			Name: "qemu",
-			Type: "raw",
-		},
-	}
-
-	diskVolume, err := virConn.LookupStorageVolByKey(volumeKey)
-	if err != nil {
-		return disk, fmt.Errorf("Can't retrieve volume %s: %v", volumeKey, err)
-	}
-	diskVolumeFile, err := diskVolume.GetPath()
-	if err != nil {
-		return disk, fmt.Errorf("Error retrieving volume file: %s", err)
-	}
-
-	disk.Source = &libvirtxml.DomainDiskSource{
-		File: &libvirtxml.DomainDiskSourceFile{
-			File: diskVolumeFile,
-		},
-	}
-
-	return disk, nil
-}
-
-func setCoreOSIgnition(d *schema.ResourceData, domainDef *libvirtxml.Domain) error {
-	if ignition, ok := d.GetOk("coreos_ignition"); ok {
-		ignitionKey, err := getIgnitionVolumeKeyFromTerraformID(ignition.(string))
-		if err != nil {
-			return err
-		}
-
-		domainDef.QEMUCommandline = &libvirtxml.DomainQEMUCommandline{
-			Args: []libvirtxml.DomainQEMUCommandlineArg{
-				{
-					Value: "-fw_cfg",
-				},
-				{
-					Value: fmt.Sprintf("name=opt/com.coreos/config,file=%s", ignitionKey),
-				},
-			},
-		}
-	}
-
-	return nil
-}
-
-func setGraphics(d *schema.ResourceData, domainDef *libvirtxml.Domain, arch string) error {
-	if arch == "s390x" || arch == "ppc64" {
-		domainDef.Devices.Graphics = nil
-		return nil
-	}
-
-	prefix := "graphics.0"
-	if _, ok := d.GetOk(prefix); ok {
-		domainDef.Devices.Graphics = []libvirtxml.DomainGraphic{{}}
-		graphicsType, ok := d.GetOk(prefix + ".type")
-		if !ok {
-			return fmt.Errorf("Missing graphics type for domain")
-		}
-
-		autoport := d.Get(prefix + ".autoport").(bool)
-		listener := libvirtxml.DomainGraphicListener{}
-
-		if listenType, ok := d.GetOk(prefix + ".listen_type"); ok {
-			switch listenType {
-			case "address":
-				listener.Address = &libvirtxml.DomainGraphicListenerAddress{}
-			case "network":
-				listener.Network = &libvirtxml.DomainGraphicListenerNetwork{}
-			case "socket":
-				listener.Socket = &libvirtxml.DomainGraphicListenerSocket{}
-			}
-		} else {
-			listenType = "none"
-		}
-
-		switch graphicsType {
-		case "spice":
-			domainDef.Devices.Graphics[0] = libvirtxml.DomainGraphic{
-				Spice: &libvirtxml.DomainGraphicSpice{},
-			}
-			domainDef.Devices.Graphics[0].Spice.AutoPort = formatBoolYesNo(autoport)
-			domainDef.Devices.Graphics[0].Spice.Listeners = []libvirtxml.DomainGraphicListener{
-				listener,
-			}
-		case "vnc":
-			domainDef.Devices.Graphics[0] = libvirtxml.DomainGraphic{
-				VNC: &libvirtxml.DomainGraphicVNC{},
-			}
-			domainDef.Devices.Graphics[0].VNC.AutoPort = formatBoolYesNo(autoport)
-			domainDef.Devices.Graphics[0].VNC.Listeners = []libvirtxml.DomainGraphicListener{
-				listener,
-			}
-		default:
-			return fmt.Errorf("This provider only supports vnc/spice as graphics type. Provided: '%s'", graphicsType)
-		}
-	}
-	return nil
-}
-
-func setCmdlineArgs(d *schema.ResourceData, domainDef *libvirtxml.Domain) {
-	var cmdlineArgs []string
-	for i := 0; i < d.Get("cmdline.#").(int); i++ {
-		for k, v := range d.Get(fmt.Sprintf("cmdline.%d", i)).(map[string]interface{}) {
-			var cmd string
-			if k == "_" {
-				// keyless cmd (eg: nosplash)
-				cmd = fmt.Sprintf("%v", v)
-			} else {
-				cmd = fmt.Sprintf("%s=%v", k, v)
-			}
-			cmdlineArgs = append(cmdlineArgs, cmd)
-		}
-	}
-	sort.Strings(cmdlineArgs)
-	domainDef.OS.Cmdline = strings.Join(cmdlineArgs, " ")
-}
-
-func setFirmware(d *schema.ResourceData, domainDef *libvirtxml.Domain) error {
-	if firmware, ok := d.GetOk("firmware"); ok {
-		firmwareFile := firmware.(string)
-		domainDef.OS.Loader = &libvirtxml.DomainLoader{
-			Path:     firmwareFile,
-			Readonly: "yes",
-			Type:     "pflash",
-			Secure:   "no",
-		}
-
-		if _, ok := d.GetOk("nvram.0"); ok {
-			nvramFile := d.Get("nvram.0.file").(string)
-			nvramTemplateFile := ""
-			if nvramTemplate, ok := d.GetOk("nvram.0.template"); ok {
-				nvramTemplateFile = nvramTemplate.(string)
-			}
-			domainDef.OS.NVRam = &libvirtxml.DomainNVRam{
-				NVRam:    nvramFile,
-				Template: nvramTemplateFile,
-			}
-		}
-	}
-
-	return nil
-}
-
-func setBootDevices(d *schema.ResourceData, domainDef *libvirtxml.Domain) {
-	for i := 0; i < d.Get("boot_device.#").(int); i++ {
-		if bootMap, ok := d.GetOk(fmt.Sprintf("boot_device.%d.dev", i)); ok {
-			for _, dev := range bootMap.([]interface{}) {
-				domainDef.OS.BootDevices = append(domainDef.OS.BootDevices,
-					libvirtxml.DomainBootDevice{
-						Dev: dev.(string),
-					})
-			}
-		}
-	}
-}
-
-func setConsoles(d *schema.ResourceData, domainDef *libvirtxml.Domain) {
-	for i := 0; i < d.Get("console.#").(int); i++ {
-		console := libvirtxml.DomainConsole{}
-		prefix := fmt.Sprintf("console.%d", i)
-		consoleTargetPortInt, err := strconv.Atoi(d.Get(prefix + ".target_port").(string))
-		if err == nil {
-			consoleTargetPort := uint(consoleTargetPortInt)
-			console.Target = &libvirtxml.DomainConsoleTarget{
-				Port: &consoleTargetPort,
-			}
-		}
-		if sourcePath, ok := d.GetOk(prefix + ".source_path"); ok {
-			console.Source = &libvirtxml.DomainChardevSource{
-				Dev: &libvirtxml.DomainChardevSourceDev{
-					Path: sourcePath.(string),
-				},
-			}
-		}
-		if targetType, ok := d.GetOk(prefix + ".target_type"); ok {
-			if console.Target == nil {
-				console.Target = &libvirtxml.DomainConsoleTarget{}
-			}
-			console.Target.Type = targetType.(string)
-		}
-		domainDef.Devices.Consoles = append(domainDef.Devices.Consoles, console)
-	}
-}
-
-func setDisks(d *schema.ResourceData, domainDef *libvirtxml.Domain, virConn *libvirt.Connect) error {
-	var scsiDisk = false
-	for i := 0; i < d.Get("disk.#").(int); i++ {
-		disk := newDefDisk(i)
-
-		prefix := fmt.Sprintf("disk.%d", i)
-		if d.Get(prefix + ".scsi").(bool) {
-			disk.Target.Bus = "scsi"
-			scsiDisk = true
-			if wwn, ok := d.GetOk(prefix + ".wwn"); ok {
-				disk.WWN = wwn.(string)
-			} else {
-				disk.WWN = randomWWN(10)
-			}
-		}
-
-		if volumeKey, ok := d.GetOk(prefix + ".volume_id"); ok {
-			diskVolume, err := virConn.LookupStorageVolByKey(volumeKey.(string))
-			if err != nil {
-				return fmt.Errorf("Can't retrieve volume %s: %v", volumeKey.(string), err)
-			}
-
-			diskVolumeName, err := diskVolume.GetName()
-			if err != nil {
-				return fmt.Errorf("Can't retrieve name for volume %s", volumeKey.(string))
-			}
-
-			diskPool, err := diskVolume.LookupPoolByVolume()
-			if err != nil {
-				return fmt.Errorf("Can't retrieve pool for volume %s", volumeKey.(string))
-			}
-
-			diskPoolName, err := diskPool.GetName()
-			if err != nil {
-				return fmt.Errorf("Can't retrieve name for pool of volume %s", volumeKey.(string))
-			}
-
-			// find out the format of the volume in order to set the appropriate
-			// driver
-			volumeDef, err := newDefVolumeFromLibvirt(diskVolume)
-			if err != nil {
-				return err
-			}
-			if volumeDef.Target != nil && volumeDef.Target.Format != nil && volumeDef.Target.Format.Type != "" {
-				if volumeDef.Target.Format.Type == "qcow2" {
-					log.Print("[DEBUG] Setting disk driver to 'qcow2' to match disk volume format")
-					disk.Driver = &libvirtxml.DomainDiskDriver{
-						Name: "qemu",
-						Type: "qcow2",
-					}
-				}
-				if volumeDef.Target.Format.Type == "raw" {
-					log.Print("[DEBUG] Setting disk driver to 'raw' to match disk volume format")
-					disk.Driver = &libvirtxml.DomainDiskDriver{
-						Name: "qemu",
-						Type: "raw",
-					}
-				}
-			} else {
-				log.Printf("[WARN] Disk volume has no format specified: %s", volumeKey.(string))
-			}
-
-			disk.Source = &libvirtxml.DomainDiskSource{
-				Volume: &libvirtxml.DomainDiskSourceVolume{
-					Pool:   diskPoolName,
-					Volume: diskVolumeName,
-				},
-			}
-		} else if rawURL, ok := d.GetOk(prefix + ".url"); ok {
-			// Support for remote, read-only http disks
-			// useful for booting CDs
-			url, err := url.Parse(rawURL.(string))
-			if err != nil {
-				return err
-			}
-
-			disk.Source = &libvirtxml.DomainDiskSource{
-				Network: &libvirtxml.DomainDiskSourceNetwork{
-					Protocol: url.Scheme,
-					Name:     url.Path,
-					Hosts: []libvirtxml.DomainDiskSourceHost{
-						{
-							Name: url.Hostname(),
-							Port: url.Port(),
-						},
-					},
-				},
-			}
-
-			if strings.HasSuffix(url.Path, ".iso") {
-				disk.Device = "cdrom"
-			}
-			if !strings.HasSuffix(url.Path, ".qcow2") {
-				disk.Driver.Type = "raw"
-			}
-		} else if file, ok := d.GetOk(prefix + ".file"); ok {
-			// support for local disks, e.g. CDs
-			disk.Source = &libvirtxml.DomainDiskSource{
-				File: &libvirtxml.DomainDiskSourceFile{
-					File: file.(string),
-				},
-			}
-
-			if strings.HasSuffix(file.(string), ".iso") {
-				disk.Device = "cdrom"
-				disk.Target = &libvirtxml.DomainDiskTarget{
-					Dev: "hda",
-					Bus: "ide",
-				}
-				disk.Driver = &libvirtxml.DomainDiskDriver{
-					Name: "qemu",
-					Type: "raw",
-				}
-			}
-		}
-
-		domainDef.Devices.Disks = append(domainDef.Devices.Disks, disk)
-	}
-
-	log.Printf("[DEBUG] scsiDisk: %t", scsiDisk)
-	if scsiDisk {
-		domainDef.Devices.Controllers = append(domainDef.Devices.Controllers,
-			libvirtxml.DomainController{
-				Type:  "scsi",
-				Model: "virtio-scsi",
-			})
-	}
-
-	return nil
-}
-
-func setFilesystems(d *schema.ResourceData, domainDef *libvirtxml.Domain) error {
-	for i := 0; i < d.Get("filesystem.#").(int); i++ {
-		fs := newFilesystemDef()
-
-		prefix := fmt.Sprintf("filesystem.%d", i)
-		if accessMode, ok := d.GetOk(prefix + ".accessmode"); ok {
-			fs.AccessMode = accessMode.(string)
-		}
-		if sourceDir, ok := d.GetOk(prefix + ".source"); ok {
-			fs.Source = &libvirtxml.DomainFilesystemSource{
-				Mount: &libvirtxml.DomainFilesystemSourceMount{
-					Dir: sourceDir.(string),
-				},
-			}
-		} else {
-			return fmt.Errorf("Filesystem entry must have a 'source' set")
-		}
-		if targetDir, ok := d.GetOk(prefix + ".target"); ok {
-			fs.Target = &libvirtxml.DomainFilesystemTarget{
-				Dir: targetDir.(string),
-			}
-		} else {
-			return fmt.Errorf("Filesystem entry must have a 'target' set")
-		}
-		if d.Get(prefix + ".readonly").(bool) {
-			fs.ReadOnly = &libvirtxml.DomainFilesystemReadOnly{}
-		} else {
-			fs.ReadOnly = nil
-		}
-
-		domainDef.Devices.Filesystems = append(domainDef.Devices.Filesystems, fs)
-	}
-	log.Printf("filesystems: %+v\n", domainDef.Devices.Filesystems)
-	return nil
-}
-
-func setCloudinit(d *schema.ResourceData, domainDef *libvirtxml.Domain, virConn *libvirt.Connect) error {
-	if cloudinit, ok := d.GetOk("cloudinit"); ok {
-		cloudinitID, err := getCloudInitVolumeKeyFromTerraformID(cloudinit.(string))
-		if err != nil {
-			return err
-		}
-		disk, err := newDiskForCloudInit(virConn, cloudinitID)
-		if err != nil {
-			return err
-		}
-		domainDef.Devices.Disks = append(domainDef.Devices.Disks, disk)
-	}
-
-	return nil
-}
-
-func setNetworkInterfaces(d *schema.ResourceData, domainDef *libvirtxml.Domain,
-	virConn *libvirt.Connect, partialNetIfaces map[string]*pendingMapping,
-	waitForLeases *[]*libvirtxml.DomainInterface) error {
-	for i := 0; i < d.Get("network_interface.#").(int); i++ {
-		prefix := fmt.Sprintf("network_interface.%d", i)
-
-		netIface := libvirtxml.DomainInterface{
-			Model: &libvirtxml.DomainInterfaceModel{
-				Type: "virtio",
-			},
-		}
-
-		// calculate the MAC address
-		var mac string
-		if macI, ok := d.GetOk(prefix + ".mac"); ok {
-			mac = strings.ToUpper(macI.(string))
-		} else {
-			var err error
-			mac, err = randomMACAddress()
-			if err != nil {
-				return fmt.Errorf("Error generating mac address: %s", err)
-			}
-		}
-		netIface.MAC = &libvirtxml.DomainInterfaceMAC{
-			Address: mac,
-		}
-
-		// this is not passed to libvirt, but used by waitForAddress
-		if waitForLease, ok := d.GetOk(prefix + ".wait_for_lease"); ok {
-			if waitForLease.(bool) {
-				*waitForLeases = append(*waitForLeases, &netIface)
-			}
-		}
-
-		// connect to the interface to the network... first, look for the network
-		if n, ok := d.GetOk(prefix + ".network_name"); ok {
-			// when using a "network_name" we do not try to do anything: we just
-			// connect to that network
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Network: &libvirtxml.DomainInterfaceSourceNetwork{
-					Network: n.(string),
-				},
-			}
-		} else if networkUUID, ok := d.GetOk(prefix + ".network_id"); ok {
-			// when using a "network_id" we are referring to a "network resource"
-			// we have defined somewhere else...
-			network, err := virConn.LookupNetworkByUUIDString(networkUUID.(string))
-			if err != nil {
-				return fmt.Errorf("Can't retrieve network ID %s", networkUUID)
-			}
-			defer network.Free()
-
-			networkName, err := network.GetName()
-			if err != nil {
-				return fmt.Errorf("Error retrieving network name: %s", err)
-			}
-			networkDef, err := getXMLNetworkDefFromLibvirt(network)
-
-			// only for DHCP, we update the host table of the network
-			if HasDHCP(networkDef) {
-				hostname := domainDef.Name
-				if hostnameI, ok := d.GetOk(prefix + ".hostname"); ok {
-					hostname = hostnameI.(string)
-				}
-				if addresses, ok := d.GetOk(prefix + ".addresses"); ok {
-					// some IP(s) provided
-					for _, addressI := range addresses.([]interface{}) {
-						address := addressI.(string)
-						ip := net.ParseIP(address)
-						if ip == nil {
-							return fmt.Errorf("Could not parse addresses '%s'", address)
-						}
-
-						log.Printf("[INFO] Adding IP/MAC/host=%s/%s/%s to %s", ip.String(), mac, hostname, networkName)
-						if err := updateOrAddHost(network, ip.String(), mac, hostname); err != nil {
-							return err
-						}
-					}
-				} else {
-					// no IPs provided: if the hostname has been provided, wait until we get an IP
-					wait := false
-					for _, iface := range *waitForLeases {
-						if iface == &netIface {
-							wait = true
-							break
-						}
-					}
-					if !wait {
-						return fmt.Errorf("Cannot map '%s': we are not waiting for DHCP lease and no IP has been provided", hostname)
-					}
-					// the resource specifies a hostname but not an IP, so we must wait until we
-					// have a valid lease and then read the IP we have been assigned, so we can
-					// do the mapping
-					log.Printf("[DEBUG] Do not have an IP for '%s' yet: will wait until DHCP provides one...", hostname)
-					partialNetIfaces[strings.ToUpper(mac)] = &pendingMapping{
-						mac:      strings.ToUpper(mac),
-						hostname: hostname,
-						network:  network,
-					}
-				}
-			}
-
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Network: &libvirtxml.DomainInterfaceSourceNetwork{
-					Network: networkName,
-				},
-			}
-		} else if bridgeNameI, ok := d.GetOk(prefix + ".bridge"); ok {
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Bridge: &libvirtxml.DomainInterfaceSourceBridge{
-					Bridge: bridgeNameI.(string),
-				},
-			}
-		} else if devI, ok := d.GetOk(prefix + ".vepa"); ok {
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Direct: &libvirtxml.DomainInterfaceSourceDirect{
-					Dev:  devI.(string),
-					Mode: "vepa",
-				},
-			}
-		} else if devI, ok := d.GetOk(prefix + ".macvtap"); ok {
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Direct: &libvirtxml.DomainInterfaceSourceDirect{
-					Dev:  devI.(string),
-					Mode: "bridge",
-				},
-			}
-		} else if devI, ok := d.GetOk(prefix + ".passthrough"); ok {
-			netIface.Source = &libvirtxml.DomainInterfaceSource{
-				Direct: &libvirtxml.DomainInterfaceSourceDirect{
-					Dev:  devI.(string),
-					Mode: "passthrough",
-				},
-			}
-		} else {
-			// no network has been specified: we are on our own
-		}
-
-		domainDef.Devices.Interfaces = append(domainDef.Devices.Interfaces, netIface)
-	}
-
-	return nil
-}
-
-func destroyDomainByUserRequest(d *schema.ResourceData, domain *libvirt.Domain) error {
-	if d.Get("running").(bool) {
-		return nil
-	}
-
-	domainID, err := domain.GetUUIDString()
-
-	if err != nil {
-		return fmt.Errorf("Error retrieving libvirt domain id: %s", err)
-	}
-
-	log.Printf("Destroying libvirt domain %s", domainID)
-	state, _, err := domain.GetState()
-	if err != nil {
-		return fmt.Errorf("Couldn't get info about domain: %s", err)
-	}
-
-	if state == libvirt.DOMAIN_RUNNING || state == libvirt.DOMAIN_PAUSED {
-		if err := domain.Destroy(); err != nil {
-			return fmt.Errorf("Couldn't destroy libvirt domain: %s", err)
 		}
 	}
