+++ libvirt/resource_libvirt_network.go
@@ -626,97 +626,3 @@ func resourceLibvirtNetworkDelete(d *schema.ResourceData, meta interface{}) erro
 	}
 	return nil
 }
-
-func waitForNetworkActive(network libvirt.Network) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		active, err := network.IsActive()
-		if err != nil {
-			return nil, "", err
-		}
-		if active {
-			return network, "ACTIVE", nil
-		}
-		return network, "BUILD", err
-	}
-}
-
-// wait for network to be up and timeout after 5 minutes.
-func waitForNetworkDestroyed(virConn *libvirt.Connect, uuid string) resource.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		log.Printf("Waiting for network %s to be destroyed", uuid)
-		network, err := virConn.LookupNetworkByUUIDString(uuid)
-		if err.(libvirt.Error).Code == libvirt.ERR_NO_NETWORK {
-			return virConn, "NOT-EXISTS", nil
-		}
-		defer network.Free()
-		return virConn, "ACTIVE", err
-	}
-}
-
-func setDhcpByCIDRAdressesSubnets(d *schema.ResourceData, networkDef *libvirtxml.Network) error {
-	if addresses, ok := d.GetOk("addresses"); ok {
-		ipsPtrsLst := []libvirtxml.NetworkIP{}
-		for _, addressI := range addresses.([]interface{}) {
-			// get the IP address entry for this subnet (with a guessed DHCP range)
-			dni, dhcp, err := setNetworkIP(addressI.(string))
-			if err != nil {
-				return err
-			}
-			if d.Get("dhcp.0.enabled").(bool) {
-				dni.DHCP = dhcp
-			} else {
-				// if a network exist with enabled but an user want to disable
-				// dhcp, we need to set dhcp struct to nil.
-				dni.DHCP = nil
-			}
-
-			ipsPtrsLst = append(ipsPtrsLst, *dni)
-		}
-		networkDef.IPs = ipsPtrsLst
-	}
-	return nil
-}
-
-func setNetworkIP(address string) (*libvirtxml.NetworkIP, *libvirtxml.NetworkDHCP, error) {
-	_, ipNet, err := net.ParseCIDR(address)
-	if err != nil {
-		return nil, nil, fmt.Errorf("Error parsing addresses definition '%s': %s", address, err)
-	}
-	ones, bits := ipNet.Mask.Size()
-	family := "ipv4"
-	if bits == (net.IPv6len * 8) {
-		family = "ipv6"
-	}
-	ipsRange := 2 ^ bits - 2 ^ ones
-	if ipsRange < 4 {
-		return nil, nil, fmt.Errorf("Netmask seems to be too strict: only %d IPs available (%s)", ipsRange-3, family)
-	}
-
-	// we should calculate the range served by DHCP. For example, for
-	// 192.168.121.0/24 we will serve 192.168.121.2 - 192.168.121.254
-	start, end := networkRange(ipNet)
-
-	// skip the .0, (for the network),
-	start[len(start)-1]++
-
-	// assign the .1 to the host interface
-	dni := &libvirtxml.NetworkIP{
-		Address: start.String(),
-		Prefix:  uint(ones),
-		Family:  family,
-	}
-
-	start[len(start)-1]++ // then skip the .1
-	end[len(end)-1]--     // and skip the .255 (for broadcast)
-
-	dhcp := &libvirtxml.NetworkDHCP{
-		Ranges: []libvirtxml.NetworkDHCPRange{
-			{
-				Start: start.String(),
-				End:   end.String(),
-			},
-		},
-	}
-
-	return dni, dhcp, nil
-}
